Move unit tests for output functionality to test_outputs.py
